fix pack system

+++ pack.py
@@ -1,10 +1,12 @@
 from pathlib import Path
 import itertools
 import shutil
+import sys
 import os
 
 PROJ_PATH = Path(__file__).parent
-BUILD_PATH = PROJ_PATH / "build" / "spruton-toys-app"
+BUILD_PATH = PROJ_PATH / "build"
+BUILD_OUT_PATH = BUILD_PATH / "spruton-toys-app"
 EXCLUDE = tuple(
     PROJ_PATH / i
     for i in (
@@ -12,11 +14,13 @@ EXCLUDE = tuple(
         "node_modules",
         ".yarn",
         ".vscode",
+        "config.yml"
     )
 )
 
-shutil.rmtree(BUILD_PATH)
-BUILD_PATH.mkdir()
+if BUILD_PATH.exists():
+    shutil.rmtree(BUILD_PATH)
+BUILD_OUT_PATH.mkdir(parents=True)
 for path in tuple(
     itertools.chain(
         PROJ_PATH.glob("*/dist/*"),
@@ -30,7 +34,7 @@ for path in tuple(
 ):
     skip = False
     for i in EXCLUDE:
-        if i in path.parents:
+        if i in itertools.chain((path, ), path.parents):
             skip = True
             break
     if skip:
@@ -39,9 +43,14 @@ for path in tuple(
     if not path.is_file():
         continue
 
-    out = BUILD_PATH / path.relative_to(PROJ_PATH)
+    out = BUILD_OUT_PATH / path.relative_to(PROJ_PATH)
     out.parent.mkdir(parents=True, exist_ok=True)
     out.touch()
     shutil.copyfile(path, out)
-os.system("del build.zip")
-os.system(f"zip -r build.zip {BUILD_PATH.absolute()}")
+
+if sys.argv[-1] == "deps":
+    os.chdir(BUILD_OUT_PATH)
+    os.system("bash pack_install_deps.sh")
+
+os.chdir(BUILD_PATH)
+os.system(f"zip -r build.zip {BUILD_OUT_PATH.relative_to(BUILD_PATH)}")
